合并Develop，准备发布2.6.0版本

+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/util/json/WxMpUserGsonAdapter.java
@@ -36,7 +36,7 @@ public class WxMpUserGsonAdapter implements JsonDeserializer<WxMpUser> {
     Integer sexId = GsonHelper.getInteger(o, "sex");
     wxMpUser.setRemark(GsonHelper.getString(o, "remark"));
     wxMpUser.setGroupId(GsonHelper.getInteger(o, "groupid"));
-    wxMpUser.setTagIds(GsonHelper.getIntArray(o, "tagid_list"));
+    wxMpUser.setTagIds(GsonHelper.getLongArray(o, "tagid_list"));
     wxMpUser.setSexId(sexId);
     if (new Integer(1).equals(sexId)) {
       wxMpUser.setSex("男");
